avr fix

+++ include/csp/interfaces/csp_if_can.h
@@ -43,7 +43,7 @@
  * The \b Begin \b flag is set for the first CAN frame in a CSP packet
  * The \b End \b flag is set for the last CAN frame in a CSP packet
  *
- * In addition to the 29 bit extended CAN header, CSP utilize four bytes in the 
+ * In addition to the 29 bit extended CAN header, CSP utilize four bytes in the
  * first CAN fragment in every CSP packet as:
  *
  * - Source:       14 bits
@@ -162,9 +162,9 @@ extern "C" {
 /**
  * Fields used to uniquely define a CSP packet within each fragment header
  */
-#define CFP2_ID_CONN_MASK ((CFP2_DST_MASK << CFP2_DST_OFFSET) | \
+#define CFP2_ID_CONN_MASK (((uint32_t)CFP2_DST_MASK << CFP2_DST_OFFSET) | \
 						   (CFP2_SENDER_MASK << CFP2_SENDER_OFFSET) | \
-					 (CFP2_PRIO_MASK << CFP2_PRIO_OFFSET) | \
+						   ((uint32_t)CFP2_PRIO_MASK << CFP2_PRIO_OFFSET) | \
 						   (CFP2_SC_MASK << CFP2_SC_OFFSET))
