+++ include/csp/interfaces/csp_if_tun.h
@@ -13,8 +13,8 @@ extern "C" {
 
 typedef struct {
 	/* Should be set before calling if_tun_init */
-	int tun_src;
-	int tun_dst;
+	uint16_t tun_src;
+	uint16_t tun_dst;
 } csp_if_tun_conf_t;
 
 void csp_if_tun_init(csp_iface_t * iface, csp_if_tun_conf_t * ifconf);
