+++ src/csp_rdp.c
@@ -502,7 +502,7 @@ bool csp_rdp_new_packet(csp_conn_t * conn, csp_packet_t * packet) {
 			csp_rdp_protocol("RDP %p: SYN-Received\n", (void *)conn);
 
 			/* Setup TX seq. */
-			unsigned int seed = csp_get_ms();
+			unsigned long seed = csp_get_ms();
 			conn->rdp.snd_iss = (uint16_t)rand_r(&seed);
 			conn->rdp.snd_nxt = conn->rdp.snd_iss + 1;
 			conn->rdp.snd_una = conn->rdp.snd_iss;
@@ -758,7 +758,7 @@ retry:
 	}
 
 	/* Randomize ISS */
-	unsigned int seed = csp_get_ms();
+	unsigned long seed = csp_get_ms();
 	conn->rdp.snd_iss = (uint16_t)rand_r(&seed);
 	conn->rdp.snd_nxt = conn->rdp.snd_iss + 1;
 	conn->rdp.snd_una = conn->rdp.snd_iss;
